Create util class for controlnet
